+++ docs/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ## [Unreleased]
 
+-   SGRString is now a dataclass not a string
+
 ## [0.8.0] - 2024-12-05
 
 ### Changed